Removed SmartProxy from UI.
- Removed Configure/SmartProxies tab and deleted any related toolbars, view yaml, product features for Smart Proxy.
- Removed SmartProxy start page shortcut from miq_shortcuts.yaml
- Cleaned up code to remove checks for miq_proxy
- Moved Tasks code into new miq_task controller that was previously in miq_proxy_controller.
- Renamed/Cleaned up tests appropriately.
- Deleted SmartProxy related tabs from Configure/Configuration, Settings accordion: Maintenance, SmartProxy, SmartProxy Affinity
(cherry picked from commit bf5b663)
