+++ vmdb/app/controllers/application_controller/ci_processing.rb
@@ -965,7 +965,7 @@ module ApplicationController::CiProcessing
     end
     @layout = session["#{self.class.session_key_prefix}_type".to_sym] if session["#{self.class.session_key_prefix}_type".to_sym]
     @current_page = @pages[:current] unless @pages.nil? # save the current page number
-    build_listnav_search_list(@view.db) if !["miq_proxy"].include?(@layout) && !session[:menu_click]
+    build_listnav_search_list(@view.db) if !["miq_task"].include?(@layout) && !session[:menu_click]
     # Came in from outside show_list partial
     unless params[:action] == "explorer"
       if params[:action] != "button" && (params[:ppsetting]  || params[:searchtag] || params[:entry] || params[:sort_choice])
