--- vmdb/app/helpers/miq_proxy_helper.rb
@@ -1,3 +1,0 @@
-module MiqProxyHelper
-  include_concern 'TextualSummary'
-end
