+++ vmdb/app/helpers/ui_constants.rb
@@ -205,8 +205,6 @@ module UiConstants
       :miqevent              => "list",
       :miqpolicy             => "list",
       :miqpolicyset          => "list",
-      :miqproxy              => "grid",
-      :miqproxybuild         => "list",
       :miqreportresult       => "list",
       :miqrequest            => "list",
       :miqtemplate           => "list",
@@ -223,7 +221,6 @@ module UiConstants
       :ms                    => "grid",
       :policy                => "list",
       :policyset             => "grid",
-      :productupdate         => "list",
       :resourcepool          => "grid",
       :service               => "grid",
       :scanhistory           => "list",
