+++ vmdb/app/presenters/menu/default_menu.rb
@@ -106,11 +106,10 @@ module Menu
 
       def configuration_menu_section
         Menu::Section.new(:set, "Configure", [
-          Menu::Item.new('configuration', N_('My Settings'),   'my_settings',  {:feature => 'my_settings', :any => true},  '/configuration/index?config_tab=ui'),
-          Menu::Item.new('my_tasks',      N_('Tasks'),         'tasks',        {:feature => 'tasks', :any => true},        '/miq_proxy/index?jobs_tab=tasks'),
-          Menu::Item.new('ops',           N_('Configuration'), 'ops_explorer', {:feature => 'ops_explorer', :any => true}, '/ops/explorer'),
-          Menu::Item.new('miq_proxy',     N_('SmartProxies'),  'miq_proxy',    {:feature => 'miq_proxy_show_list'},        '/miq_proxy'),
-          Menu::Item.new('about',         N_('About'),         'about',        {:feature => 'about'},                      '/support/index?support_tab=about')
+          Menu::Item.new('configuration', _('My Settings'),   'my_settings',  {:feature => 'my_settings', :any => true},  '/configuration/index?config_tab=ui'),
+          Menu::Item.new('my_tasks',      _('Tasks'),         'tasks',        {:feature => 'tasks', :any => true},        '/miq_task/index?jobs_tab=tasks'),
+          Menu::Item.new('ops',           _('Configuration'), 'ops_explorer', {:feature => 'ops_explorer', :any => true}, '/ops/explorer'),
+          Menu::Item.new('about',         _('About'),         'about',        {:feature => 'about'},                      '/support/index?support_tab=about')
         ])
       end
