+++ vmdb/app/views/layouts/_tasks.html.haml
@@ -1,3 +1,3 @@
 #main_div
-  = render :partial => 'miq_proxy/tasks_options'
+  = render :partial => 'miq_task/tasks_options'
   = render :partial => 'layouts/gtl'
