--- vmdb/app/views/layouts/listnav/_miq_proxy.html.haml
@@ -1,61 +1,0 @@
-- if @record.try(:name)
-  #accordion.panel-group
-    = patternfly_accordion_panel(truncate(@record.name, :length => truncate_length), true, "icon") do
-      = render :partial => "layouts/quadicon", :locals => {:mode => :icon, :item => @record, :size => 72, :typ => :listnav}
-
-    = patternfly_accordion_panel(_("Properties"), false, "miqproxy_prop") do
-      %ul.nav.nav-pills.nav-stacked
-        %li
-          = link_to(_('Summary'),
-            {:action => 'show', :id => @record, :display => 'main'},
-            :title => _("Show Summary"))
-        %li
-          = link_to(_('Log Viewer'),
-            {:action => 'log_viewer', :id => @record},
-            :title => _("View the SmartProxy log"))
-
-    = patternfly_accordion_panel(_("Relationships"), false, "miqproxy_rel") do
-      %ul.nav.nav-pills.nav-stacked
-        - if role_allows(:feature => "host_show")
-          - if @record.host.nil?
-            %li.disabled
-              = link_to(_("Installed on Host: None"), "#")
-          - else
-            %li
-              = link_to(_("Installed on Host: %s") % @record.host.name,
-                {:controller => "host", :action => 'show', :id => @record.host.id.to_s},
-                :title => _("Show the Host this SmartProxy is installed on"))
-
-        - if role_allows(:feature => "ems_infra_show") && !@record.ext_management_system.nil?
-          %li
-            = link_to("#{ui_lookup(:table => "ems_infra")}: #{@record.ext_management_system.name}",
-              {:controller => "ems_infra", :action => 'show', :id => @record.ext_management_system.id.to_s},
-              {:title => _("Show this SmartProxy's %s") % ui_lookup(:table => "ems_infra")},
-              '/images/icons/16/link_external.gif')
-
-        - if role_allows(:feature => "storage_show_list")
-          = li_link_if_nonzero(:count => @record.storages.length,
-            :record_id                => @record.id,
-            :tables                   => 'storages',
-            :display                  => 'storages',
-            :action                   => 'show')
-
-        - if role_allows(:feature => "vm_show_list")
-          - if @record.vms.length == 0
-            %li.disabled
-              = link_to(_("Managed VMs (%s)") % @record.vms.length, "#")
-          - else
-            %li
-              = link_to(_("Managed VMs (%s)") % @record.vms.length,
-                {:action => 'show', :id => @record, :display => 'vms'},
-                :title => _("Show all managed VMs"))
-
-        - if role_allows(:feature => "miq_template_show_list")
-          - if @record.miq_templates.length == 0
-            %li.disabled
-              = link_to(_("Templates (%s)") % @record.miq_templates.length, "#")
-          - else
-            %li
-              = link_to(_("Templates (%s)") % @record.miq_templates.length,
-                {:action => 'show', :id => @record, :display => 'miq_templates'},
-                :title => _("Show all Templates"))
