--- vmdb/app/views/miq_proxy/_form.html.haml
@@ -1,93 +1,0 @@
-- url = url_for(:action => 'form_field_changed', :id => "#{@miq_proxy.id || 'new'}")
-/ [form:miq_proxy]
-#form_div
-  = render :partial => "layouts/flash_msg"
-  - if %w(new create).include?(controller.action_name)
-    %h3=_("Select SmartProxy Host")
-    %table.style1
-      %tr
-        %td.key=_("Name (from chosen Host)")
-        %td.wide{:colspan => "2"}
-          = select_tag("host",
-            options_for_select([["<#{_('Choose')}>", "<Choose>"]] + Array(@edit[:avail_hosts].invert).sort_by { |name, _id| name.downcase }, @edit[:host]),
-            "data-miq_observe" => {:url => url}.to_json)
-  - else
-    %h3=_("Settings")
-    %table.style1
-      %tr
-        %td.key=_("Heartbeat Frequency")
-        %td{:width => "70"}
-          = select_tag("hb_freq_mins",
-            options_for_select((0..60).to_a, @edit[:new][:hb_freq_mins]),
-            "data-miq_observe" => {:url => url}.to_json)
-        %td
-          = select_tag("hb_freq_secs",
-            options_for_select(0.step(55, 5).to_a, @edit[:new][:hb_freq_secs]),
-            "data-miq_observe" => {:url => url}.to_json)
-
-      - if false
-        - # Not in use, hiding for now - Sprint 41
-        %tr
-          %td.key=_("Self Update Frequency")
-          %td{:width => "70"}
-            = select_tag("update_freq_days",
-              options_for_select((0..14).to_a, @edit[:new][:update_freq_days]),
-              "data-miq_observe" => {:url => url}.to_json)
-          %td
-            = select_tag("update_freq_hours",
-              options_for_select((0..23).to_a, @edit[:new][:update_freq_hours]),
-              "data-miq_observe" => {:url => url}.to_json)
-
-      %tr
-        %td.key=_("Read Only Mode")
-        %td{:colspan => 2}
-          = check_box_tag("readonly", "1", @edit[:new][:readonly],
-          "data-miq_observe_checkbox" => {:url => url}.to_json)
-      %tr
-        %td.key=_("Web Services Listen Port")
-        %td.wide{:colspan => "2"}
-          = text_field_tag("ws_port", @edit[:new][:ws_port],
-          :maxlength         => 6,
-          "data-miq_observe" => {:interval => '.5', :url => url}.to_json)
-          =_("(max 6 digits)")
-      %tr
-        %td.key=_("Log Level")
-        %td.wide{:colspan => "2"}
-          = select_tag("log_level",
-            options_for_select(%w(fatal error warn info debug), @edit[:new][:log_level]),
-            "data-miq_observe" => {:url => url}.to_json)
-      %tr
-        %td.key=_("Log Wrap Size")
-        %td.wide{:colspan => "2"}
-          = text_field_tag("log_wrapsize", @edit[:new][:log_wrapsize],
-            :maxlength         => 3,
-            "data-miq_observe" => {:interval => '.5', :url => url}.to_json)
-          (1-999 MB)
-      %tr
-        %td.key=_("Log Wrap Time")
-        %td{:width => "70"}
-          = select_tag("log_wraptime_days",
-            options_for_select((1..30).to_a, @edit[:new][:log_wraptime_days]),
-            "data-miq_observe" => {:url => url}.to_json)
-        %td
-          = select_tag("log_wraptime_hours",
-            options_for_select((0..23).to_a, @edit[:new][:log_wraptime_hours]),
-            "data-miq_observe" => {:url => url}.to_json)
-
-  - if %w(new create).include?(controller.action_name)
-    %table{:width => "100%"}
-      %tr
-        %td{:align => "right"}
-          #buttons_on
-            = button_tag(_('Add'),
-              :class   => "btn btn-primary",
-              :alt     => t = _("Add this SmartProxy"),
-              :title   => t,
-              :onclick => "miqAjaxButton('#{url_for(:action => 'create', :id => 'new', :button => 'add')}');")
-            = button_tag(t = _('Cancel'),
-              :class   => "btn btn-default",
-              :alt     => t,
-              :title   => t,
-              :onclick => "miqAjaxButton('#{url_for(:action => 'create', :id => 'new', :button => 'cancel')}');")
-  - else
-    = render :partial => '/layouts/edit_form_buttons', :locals  => {:record_id => @miq_proxy.id, :ajax_buttons => true}
