--- vmdb/app/views/miq_proxy/_main.html.haml
@@ -1,7 +1,0 @@
-= render :partial => "layouts/flash_msg"
-.row
-  .col-sm-12.col-md-12.col-lg-6
-    = render :partial => "shared/summary/textual", :locals => {:title => _("Properties"), :items => textual_group_properties}
-    = render :partial => "shared/summary/textual", :locals => {:title => _("Relationships"), :items => textual_group_relationships}
-  .col-sm-12.col-md-12.col-lg-6
-    = render :partial => "shared/summary/textual_tags", :locals => {:title => _("Smart Management"), :items => textual_group_smart_management}
