--- vmdb/app/views/miq_proxy/edit.html.haml
@@ -1,1 +1,0 @@
-= render :partial => 'form'
