--- vmdb/app/views/miq_proxy/jobs.html.haml
@@ -1,1 +1,0 @@
-= render :partial => "layouts/tasks"
