--- vmdb/app/views/miq_proxy/show_list.html.haml
@@ -1,2 +1,0 @@
-#main_div
-  = render :partial => 'layouts/gtl'
