+++ vmdb/app/views/miq_task/_tasks_options.html.haml
@@ -93,3 +93,4 @@
               :remote                => true,
               :title                 => t)
   %hr
+
