+++ vmdb/app/views/miq_task/jobs.html.haml
@@ -1,0 +1,1 @@
+= render :partial => "layouts/tasks"
