--- vmdb/app/views/ops/_settings_smartproxy_affinity_tab.html.haml
@@ -1,4 +1,0 @@
-- url = url_for(:action =>'settings_form_field_changed', :id=>"#{@sb[:active_tab].split('_').last}")
-- if @sb[:active_tab] == "settings_smartproxy_affinity"
-  = render :partial => "layouts/flash_msg"
-  = render :partial => "smartproxy_affinity"
