--- vmdb/product/views/MiqProxy.yaml
@@ -1,78 +1,0 @@
-#
-# This is an MIQ Report configuration file
-#   Single value parameters are specified as:
-#     single_value_parm: value
-#   Multiple value parameters are specified as:
-#     multi_value_parm:
-#       - value 1
-#       - value 2
-#
-
-# Report title
-title: SmartProxy
-
-# Menu name
-name: SmartProxy
-
-# Main DB table report is based on
-db: MiqProxy
-
-# Columns to fetch from the main table
-cols:
-- name
-- last_heartbeat
-- version
-- v_arch
-- ws_port
-- region_description
-
-# Included tables (joined, has_one, has_many) and columns
-include:
-
-# Order of columns (from all tables)
-col_order:
-- name
-- last_heartbeat
-- version
-- v_arch
-- ws_port
-- region_description
-
-# Column titles, in order
-headers:
-- Name
-- Last Heartbeat
-- Version
-- Architecture
-- WS Port
-- Region
-
-# Condition(s) string for the SQL query
-conditions:
-
-# Order string for the SQL query
-order: Ascending
-
-# Columns to sort the report on, in order
-sortby:
-- name
-
-# Group rows (y=yes,n=no,c=count)
-group: n
-
-# Graph type
-#   Bar
-#   Column
-#   ColumnThreed
-#   ParallelThreedColumn
-#   Pie
-#   PieThreed
-#   StackedBar
-#   StackedColumn
-#   StackedThreedColumn
-
-graph: 
-
-# Dimensions of graph (1 or 2)
-#   Note: specifying 2 for a single dimension graph may not return expected results
-dims:
