+++ vmdb/spec/controllers/miq_task_controller_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-describe MiqProxyController do
+describe MiqTaskController do
 
   context "#tasks_condition" do
     subject { controller.send(:tasks_condition, @opts) }
@@ -780,59 +780,6 @@ describe MiqProxyController do
 
     def set_opts(hsh)
       hsh.each_pair { |k,v| @opts[k] = v }
-    end
-  end
-
-  context "#button" do
-    it "when VM Right Size Recommendations is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "vm_right_size"})
-      controller.should_receive(:vm_right_size)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when VM Migrate is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "vm_migrate"})
-      controller.instance_variable_set(:@refresh_partial,"layouts/gtl")
-      controller.should_receive(:prov_redirect).with("migrate")
-      controller.should_receive(:render)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when VM Retire is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "vm_retire"})
-      controller.should_receive(:retirevms).once
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when VM Manage Policies is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "vm_protect"})
-      controller.should_receive(:assign_policies).with(VmOrTemplate)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when MiqTemplate Manage Policies is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "miq_template_protect"})
-      controller.should_receive(:assign_policies).with(VmOrTemplate)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when VM Tag is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "vm_tag"})
-      controller.should_receive(:tag).with(VmOrTemplate)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
-    end
-
-    it "when MiqTemplate Tag is pressed" do
-      controller.instance_variable_set(:@_params, {:pressed => "miq_template_tag"})
-      controller.should_receive(:tag).with(VmOrTemplate)
-      controller.button
-      controller.send(:flash_errors?).should_not be_true
     end
   end
 end
